+++ bboss-elasticsearch-rest/src/main/java/org/frameworkset/elasticsearch/ElasticSearchHelper.java
@@ -143,7 +143,7 @@ public class ElasticSearchHelper {
 			elasticsearchPropes.put("elasticsearch.sliceScrollBlockedWaitTimeout",
 					ElasticSearchHelper._getStringValue(serverName,"elasticsearch.sliceScrollBlockedWaitTimeout",configContext,"0"));
 			elasticsearchPropes.put("elasticsearch.includeTypeName",
-					ElasticSearchHelper._getStringValue(serverName,"elasticsearch.includeTypeName",configContext,"true"));
+					ElasticSearchHelper._getStringValue(serverName,"elasticsearch.includeTypeName",configContext,"false"));
 			elasticsearchPropes.put("elasticsearch.scrollThreadCount",
 					ElasticSearchHelper._getStringValue(serverName,"elasticsearch.scrollThreadCount",configContext,"200"));
 			elasticsearchPropes.put("elasticsearch.scrollThreadQueue",