+++ docs/changelog.md
@@ -16,6 +16,12 @@ https://esdoc.bbossgroups.com/#/development
 6. 提供快速而高效的数据导入ES工具
 7. APM开源产品pinpoint官方Elasticsearch bboss 客户端性能监控插件，插件地址： https://github.com/naver/pinpoint/tree/master/plugins/elasticsearch-bboss
 
+# v5.8.9 功能改进
+
+1.改进检索Meta数据：增加seqNo和primaryTerm属性。
+
+2.includeTypeName配置默认设置为false
+
 # v5.8.8 功能改进
 
 1.改进检索Meta数据：增加explanation属性。